fix: logo position
