+++ mrtt-ui/src/components/Header/Header.js
@@ -4,7 +4,7 @@ import theme from '../../styles/theme'
 import HeaderMenu from './HeaderMenu'
 import { ReactComponent as GmwLogo } from '../../assets/gmw-logo.svg'
 import { useFeatureFlags } from '../../hooks/useFeatureFlags'
-import { Header as HeaderV2, Logo } from '../../styles/v2/containers/header'
+import { Header as HeaderV2 } from '../../styles/v2/containers/header'
 
 const HeaderContainer = styled('header')`
   background-color: ${theme.color.secondary};
@@ -31,11 +31,7 @@ const Header = () => {
   const { showNewLandingPage } = useFeatureFlags()
 
   if (showNewLandingPage) {
-    return (
-      <HeaderV2>
-        <Logo src='/images/landing/logo.webp' />
-      </HeaderV2>
-    )
+    return <HeaderV2></HeaderV2>
   }
 
   return (
